backend-django/requirements.txt
mysqlclient==2.2.1
django==4.2.9
djangorestframework==3.14.0
django-filter==23.5
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
dj_database_url==2.1.0
gunicorn==20.1.0
gevent==22.10.2
uvicorn==0.21.0
